11.21 khc

+++ app/src/main/java/com/example/kk/arttraining/ui/valuation/view/ValuationMain.java
@@ -325,7 +325,7 @@ public class ValuationMain extends BaseActivity implements IValuationMain {
                     teacherGridViewAdapter = new ValuationGridViewAdapter(this, teacherList);
                     valuationGvTeacher.setAdapter(teacherGridViewAdapter);
                     valuationGvTeacher.setOnItemClickListener(new ChooseTeacherItemClick());
-                    int price = 0;
+                    float price = 0;
                     for (int i = 0; i < teacherList.size(); i++) {
                         TecInfoBean tecInfoBean = teacherList.get(i);
                         price = price + tecInfoBean.getAss_pay();
